tests: fix some flakiness with e2e tests (#2872)
Also, contains a small refactor of project settings container components (refactored to `.tsx`)

---------
