+++ tests/cypress.config.ts
@@ -1,20 +1,19 @@
 import { defineConfig } from "cypress";
 
 export default defineConfig({
+  e2e: {
+    baseUrl: "http://localhost:3000",
+    specPattern: "cypress/e2e/**/*.{js,jsx,ts,tsx}",
+  },
   env: {
     USE_FIXTURES: true,
     CLOUD_STORAGE_READWRITE_ENABLED: false,
   },
-
+  retries: {
+    runMode: 2,
+    openMode: 0,
+  },
   video: false,
   viewportWidth: 1600,
   viewportHeight: 1200,
-
-  e2e: {
-    // setupNodeEvents(on, config) {
-    //   return require("./cypress/plugins/index.ts")(on, config);
-    // },
-    baseUrl: "http://localhost:3000",
-    specPattern: "cypress/e2e/**/*.{js,jsx,ts,tsx}",
-  },
 });
