+++ tests/cypress/e2e/dashboard.spec.ts
@@ -71,6 +71,7 @@ describe("dashboard", () => {
   it("user does not have own project but has visited projects", () => {
     fixtures
       .projects()
+      .sessionServersEmpty()
       .entitySearch("getEntities", "kgSearch/emptySearch.json", "0")
       .getLastVisitedProjects();
     const files = {
@@ -119,6 +120,7 @@ describe("dashboard", () => {
   it("user has own projects and recently visited projects", () => {
     fixtures
       .projects()
+      .sessionServersEmpty()
       .entitySearch("getEntities", "kgSearch/search.json", "7")
       .getLastVisitedProjects(
         "getLastVisitedProjects",
